# sdks/sdkwork-missory-app-sdk/openapi/sdkwork-missory-app-api.sdkgen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDKWork Missory App API",
    "version": "0.1.0",
    "description": "Missory (念忆) app-api surface: people, relationships, memories, stories, reminders, and the draft-only social AI assistant. Authority contract for the `sdkwork-missory-app-api` API authority mounted under the locked app-api prefix."
  },
  "servers": [
    {
      "url": "https://api.sdkwork.com",
      "description": "SDKWork cloud API"
    },
    {
      "url": "http://127.0.0.1:8460",
      "description": "Local standalone gateway"
    }
  ],
...
  "paths": {
...
    "/app/v3/api/missory/memories/extract": {
      "post": {
        "operationId": "memories.extract",
        "tags": [
          "missory"
        ],
        "summary": "Extract candidate memories from free text (AI assist, candidates only).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissoryMemoryExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memory page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissoryMemoryPageResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "RFC 9457 problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-owner": "sdkwork-missory",
        "x-sdkwork-api-authority": "sdkwork-missory-app-api",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-request-context": "MissoryRequestContext",
        "x-sdkwork-domain": "communication",
        "x-sdkwork-resource": "memories",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "owner",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Int64Id": {
        "type": "string",
        "format": "int64",
        "pattern": "^-?[0-9]+$",
        "x-sdkwork-int64-string": true
      },
      "ProblemDetail": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "traceId": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
...
      "MissoryImportance": {
        "type": "string",
        "enum": [
          "low",
          "normal",
          "high",
          "core"
        ]
      },
      "MissoryMemoryType": {
        "type": "string",
        "enum": [
          "semantic",
          "episodic",
          "temporal",
          "relationship",
          "preference",
          "commitment"
        ]
      },
      "MissoryMemoryStatus": {
        "type": "string",
        "enum": [
          "candidate",
          "confirmed",
          "rejected",
          "archived",
          "expired"
        ]
      },
      "MissoryMemoryOrigin": {
        "type": "string",
        "enum": [
          "fact",
          "inference"
        ]
      },
      "MissorySourceKind": {
        "type": "string",
        "enum": [
          "user-input",
          "file",
          "conversation",
          "ai-extraction",
          "ai-inference",
          "system"
        ]
      },
...
      "MissoryMemory": {
        "type": "object",
        "required": [
          "id",
          "personId",
          "type",
          "content",
          "origin",
          "status",
          "sourceKind",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Int64Id"
          },
          "personId": {
            "$ref": "#/components/schemas/Int64Id"
          },
          "storyId": {
            "$ref": "#/components/schemas/Int64Id"
          },
          "type": {
            "$ref": "#/components/schemas/MissoryMemoryType"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/MissoryMemoryOrigin"
          },
          "status": {
            "$ref": "#/components/schemas/MissoryMemoryStatus"
          },
          "confidence": {
            "type": "number",
            "description": "Inference-only confidence in [0,1]."
          },
          "sourceReason": {
            "type": "string",
            "description": "Inference-only reasoning note."
          },
          "sourceKind": {
            "$ref": "#/components/schemas/MissorySourceKind"
          },
          "sourceRef": {
            "type": "string"
          },
          "importance": {
            "$ref": "#/components/schemas/MissoryImportance"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MissoryMemoryExtractRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "text"
        ],
        "properties": {
          "personId": {
            "$ref": "#/components/schemas/Int64Id",
            "description": "Optional target person; omitted resolves by name mentions in text."
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalItems": {
            "type": "string"
          },
          "totalPages": {
            "type": "integer"
          },
          "nextCursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "MissoryMemoryPageResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "items",
                  "pageInfo"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MissoryMemory"
                    }
                  },
                  "pageInfo": {
                    "$ref": "#/components/schemas/PageInfo"
                  }
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    }
  },
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ]
}